% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_dunn_letter.R
\name{add_dunn_letters}
\alias{add_dunn_letters}
\title{add significance}
\usage{
add_dunn_letters(data, variable)
}
\arguments{
\item{data}{data frame}

\item{variable}{variable}
}
\value{

}
\description{
add significance
}
\examples{
#' library(ggplot2)
data <- iris
colnames(data)[5] <- "group"
bar_plot(data = data, variable = "Petal.Width") + add_dunn_letters(data = data, variable = "Petal.Width")
}
\author{
Zhonghui Gai
}
