% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_signif.R
\name{add_signif}
\alias{add_signif}
\title{add significance}
\usage{
add_signif(
  comparisons = NULL,
  test = "wilcox.test",
  y_position = NULL,
  map_signif_level = TRUE,
  step_increase = 0,
  tip_length = 0,
  ...
)
}
\arguments{
\item{comparisons}{a list}

\item{test}{the type of test}

\item{y_position}{numeric}

\item{map_signif_level}{logical}

\item{step_increase}{numeric}

\item{tip_length}{numeric}

\item{...}{other parameters}
}
\value{

}
\description{
add significance
}
\examples{
library(ggplot2)
data <- iris
colnames(data)[5] <- "group"
bar_plot(data = data, variable = "Petal.Width") + add_signif(list(c("setosa", "versicolor")))
}
\author{
Zhonghui Gai
}
