% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/barplot.R
\name{bar_plot}
\alias{bar_plot}
\title{barplot using grouping information}
\usage{
bar_plot(
  data,
  variable = NULL,
  jitter = FALSE,
  border = TRUE,
  fun.data = "mean_sdl",
  angle = 45
)
}
\arguments{
\item{data}{`data.frame` the import data}

\item{variable}{`character`, the variable to be ploted}

\item{jitter}{`logical`, point jitter or not}

\item{border}{`logical`, whether to show the plot border}

\item{fun.data}{`character`, the type of error bar, one of "mean_sdl" or "mean_se",
the default value is "mean_sdl"}

\item{angle}{`numeric`, the x axis test angle}
}
\value{
a ggplots object
}
\description{
barplot using grouping information
}
\examples{
library(ggplot2)
data <- iris
colnames(data)[5] <- "group"
bar_plot(data = data, variable = "Petal.Width")
}
