% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_class.R
\name{check_class}
\alias{check_class}
\title{Check if a data frame having grouping information, the the class of the rest vvariables are numeric!}
\usage{
check_class(data)
}
\arguments{
\item{data}{a data frame}
}
\value{
a logical value
}
\description{
Check if a data frame having grouping information, the the class of the rest vvariables are numeric!
}
\examples{
check_class(iris)
data <- iris
colnames(data)[5] <- "group"
check_class(data)
}
\author{
Zhonghui Gai
}
