% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dunn_letters.R
\name{dunn_letters}
\alias{dunn_letters}
\title{A function for non-parametric multi-group test using dunn.test package}
\usage{
dunn_letters(data, letter = TRUE, adjust = FALSE)
}
\arguments{
\item{data}{a data frame with grouping information}

\item{letter}{show p value using letters, default value is TRUE}

\item{adjust}{adjust p value using "bh", default value is FALSE}
}
\value{
a data frame
}
\description{
A function for non-parametric multi-group test using dunn.test package
}
\examples{
data <- iris
colnames(data)[5] <- "group"
dunn_letters(data = data, adjust = TRUE)
dunn_letters(data = data, letter = FALSE)
}
\author{
Zhonghui Gai
}
