% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aov_letters.R
\name{gmax}
\alias{gmax}
\title{Calculate the max values among different groups}
\usage{
gmax(data)
}
\arguments{
\item{data}{a data frame}
}
\value{

}
\description{
Calculate the max values among different groups
}
