% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggpub.R
\name{setPath}
\alias{setPath}
\alias{setDpi}
\alias{setType}
\title{Change defaults}
\usage{
setPath(option)

setDpi(option)

setType(option)
}
\arguments{
\item{option}{new value for default}
}
\description{
Functions to change the defaults for ggpub.
}
\examples{
\dontrun{
setPath('my/new/path')
showGgpubDefaults() # Displays new default.
}
}
