% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/violinplot.R
\name{violin_plot}
\alias{violin_plot}
\title{violinplot using grouping information}
\usage{
violin_plot(data, variable = NULL, border = TRUE, angle = 45)
}
\arguments{
\item{data}{`data.frame` the import data}

\item{variable}{`character`, the variable to be ploted}

\item{border}{`logical`, whether to show the plot border}

\item{angle}{`numeric`, the x axis test angle}
}
\value{

}
\description{
violinplot using grouping information
}
\examples{
library(ggplot2)
data <- iris
colnames(data)[5] <- "group"
violin_plot(data = data, variable = "Petal.Width", angle = 0) + add_dunn_letters(data = data, variable = "Petal.Width")
violin_plot(data = data, variable = "Petal.Width", angle = 0) + add_aov_letters(data = data, variable = "Petal.Width")
violin_plot(data = data, variable = "Petal.Width", angle = 0) + add_signif(list(c("setosa", "versicolor")))
}
\author{
Zhonghui Gai
}
