% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Themes.R
\name{theme_pub}
\alias{theme_pub}
\title{Theme to make creating journal ready figures easier}
\usage{
theme_pub(base_size = 11, base_family = "Arial", plot_grid = "none",
  lines_lwd = 0.5, bg_col = "white", axis_lines = TRUE,
  horz_grid = ifelse(plot_grid == "none", FALSE, TRUE),
  vert_grid = ifelse(plot_grid == "none", FALSE, TRUE),
  ticks_type = "outer", ticks_xy = "xy", alpha_leg = 0.1,
  bord_size = lines_lwd, legend_bg = "white", strip_bg = "white",
  title_size = base_size * 1.2, legend_size = base_size * 0.9,
  grid_thick = 1, grid_type = "solid", grid_cols = c("grey10", "grey30"),
  legend_title = TRUE, plot_box = TRUE)
}
\arguments{
\item{base_size}{Font size}

\item{base_family}{Font used for everything, unless axis fonts etc. are defined}

\item{plot_grid}{= Do you want a grid? (all = major/minor, none = no grid, major = major grid)}

\item{lines_lwd}{= width of lines}

\item{bg_col}{= background colour}

\item{horz_grid}{= Do you want a horizontal grid?}

\item{ticks_xy}{= Do you want ticks on the x or y axis? "x" = x-axis only, "y" = y-axis only, "xy" = both axes.}

\item{alpha_leg}{= opacity of the legend. 0 = totally transparent}

\item{bord_size}{= width of a rectangular border}

\item{strip_bg}{= colour background for facets}

\item{title_size}{= font size of title}

\item{legend_size}{= font size of legend}

\item{grid_thick}{= A multiplier to apply to the grid lines.}

\item{grid_type}{= Grid type. Default is a solid line}

\item{grid_cols}{= Colour of the grid. 2 element vector. First element is major grid colour. If only one element, the first will be used for minor grid.}

\item{legend_title}{= Do you want the legend title? (TRUE/FALSE)}

\item{title_font}{= font use for title}

\item{base_col}{= font colour}
}
\value{
ggplot2 theme
}
\description{
Theme to make creating journal ready figures easier
}
